Move Email and Marketing tab components to there own subfolder
